Continued WIP
There are a number of TODOs outstanding, but basically pyprima has been
integrated into the prima package.

A besdies the TODOs, the pyprima tests need to set 'fortran': False and
we may want to consolidate some tests and parameterize them to use
fortran false/true

+++ python/_prima.cpp
@@ -166,7 +166,7 @@ PYBIND11_MODULE(_prima, m) {
         // The following are not options, but part of the problem. We are using the options dictionary
         // as a convenient way to pass them to C++.
         if(options_dict.contains("f0"))        { problem.f0        = options_dict["f0"].cast<double>(); }
-        if(options_dict.contains("m_nlcon"))   { problem.m_nlcon   = options_dict["m_nlcon"].cast<double>(); }
+        if(options_dict.contains("m_nlcon"))   { problem.m_nlcon   = options_dict["m_nlcon"].cast<int>(); }
         if(options_dict.contains("nlconstr0"))    {
           try
           {
